feat: FSD 2.0 아키텍처 적용, 파일 이관작업
