+++ index.html
@@ -6,13 +6,6 @@
     <title>Chapter 2-3. 관심사 분리와 폴더구조</title>
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <script src="https://cdn.tailwindcss.com"></script>
-    <style>
-      html,
-      body {
-        background: #fff;
-        color: #000;
-      }
-    </style>
   </head>
 
   <body>
