+++ src/app/styles/index.css
@@ -1,0 +1,5 @@
+html,
+body {
+  background: #fff;
+  color: #000;
+}
