--- src/components/Input.tsx
@@ -1,18 +1,0 @@
-import type { ComponentPropsWithRef } from "react"
-
-import { mergeClasses } from "../utils/classUtils"
-
-type InputProps = ComponentPropsWithRef<"input">
-
-export function Input({ className, type, ...rest }: InputProps) {
-  return (
-    <input
-      type={type}
-      className={mergeClasses(
-        "border-input ring-offset-background placeholder:text-muted-foreground focus-visible:ring-ring flex h-10 w-full rounded-md border bg-white px-3 py-2 text-sm file:border-0 file:bg-transparent file:text-sm file:font-medium focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50",
-        className,
-      )}
-      {...rest}
-    />
-  )
-}
