+++ src/features/comment-management/hooks/useComments.ts
@@ -1,0 +1,97 @@
+import { useState } from "react"
+
+import { commentApi } from "@/entities/comment/api"
+import type { Comment } from "@/entities/comment/model"
+
+export const useComments = () => {
+  const [comments, setComments] = useState<Record<number, Comment[]>>({})
+  const [selectedComment, setSelectedComment] = useState<Comment | null>(null)
+  const [newComment, setNewComment] = useState({ body: "", postId: null as number | null, userId: 1 })
+  const [showAddCommentDialog, setShowAddCommentDialog] = useState(false)
+  const [showEditCommentDialog, setShowEditCommentDialog] = useState(false)
+
+  const fetchComments = async (postId: number) => {
+    if (comments[postId]) return
+    try {
+      const data = await commentApi.getCommentsByPostId(postId)
+      setComments((prev) => ({ ...prev, [postId]: data.comments }))
+    } catch (error) {
+      console.error("댓글 가져오기 오류:", error)
+    }
+  }
+
+  const addComment = async () => {
+    if (!newComment.postId) return
+    try {
+      const data = await commentApi.createComment(newComment)
+      setComments((prev) => ({
+        ...prev,
+        [data.postId]: [...(prev[data.postId] || []), data],
+      }))
+      setShowAddCommentDialog(false)
+      setNewComment({ body: "", postId: null, userId: 1 })
+    } catch (error) {
+      console.error("댓글 추가 오류:", error)
+    }
+  }
+
+  const updateComment = async () => {
+    if (!selectedComment) return
+    try {
+      const data = await commentApi.updateComment(selectedComment.id, selectedComment.body)
+      setComments((prev) => ({
+        ...prev,
+        [data.postId]: prev[data.postId].map((comment) => (comment.id === data.id ? data : comment)),
+      }))
+      setShowEditCommentDialog(false)
+    } catch (error) {
+      console.error("댓글 업데이트 오류:", error)
+    }
+  }
+
+  const deleteComment = async (id: number, postId: number) => {
+    try {
+      await commentApi.deleteComment(id)
+      setComments((prev) => ({
+        ...prev,
+        [postId]: prev[postId].filter((comment) => comment.id !== id),
+      }))
+    } catch (error) {
+      console.error("댓글 삭제 오류:", error)
+    }
+  }
+
+  const likeComment = async (id: number, postId: number) => {
+    try {
+      const currentComment = comments[postId].find((c) => c.id === id)
+      if (!currentComment) return
+
+      const data = await commentApi.likeComment(id, currentComment.likes + 1)
+      setComments((prev) => ({
+        ...prev,
+        [postId]: prev[postId].map((comment) =>
+          comment.id === data.id ? { ...data, likes: comment.likes + 1 } : comment,
+        ),
+      }))
+    } catch (error) {
+      console.error("댓글 좋아요 오류:", error)
+    }
+  }
+
+  return {
+    comments,
+    selectedComment,
+    setSelectedComment,
+    newComment,
+    setNewComment,
+    showAddCommentDialog,
+    setShowAddCommentDialog,
+    showEditCommentDialog,
+    setShowEditCommentDialog,
+    fetchComments,
+    addComment,
+    updateComment,
+    deleteComment,
+    likeComment,
+  }
+}
