+++ src/main.tsx
@@ -1,7 +1,10 @@
+import "@/app/styles/index.css"
+
 import { StrictMode } from "react"
 import { createRoot } from "react-dom/client"
+import { RouterProvider } from "react-router-dom"
 
-import { App } from "./App"
+import { appRouter } from "@/app/router"
 
 const rootElement = document.getElementById("root")
 
@@ -11,6 +14,6 @@ if (!rootElement) {
 
 createRoot(rootElement).render(
   <StrictMode>
-    <App />
+    <RouterProvider router={appRouter} />
   </StrictMode>,
 )
