+++ src/pages/posts-manager/ui/PostsManagerPage.tsx
@@ -1,0 +1,5 @@
+import { PostsManagerWidget } from "@/widgets/posts-manager/ui"
+
+export function PostsManagerPage() {
+  return <PostsManagerWidget />
+}
