+++ src/shared/api/index.ts
@@ -1,0 +1,2 @@
+export * from "./base"
+export type * from "./types"
