+++ src/shared/api/types.ts
@@ -1,0 +1,18 @@
+export interface ApiResponse<T> {
+  data: T
+  status: number
+  message?: string
+}
+
+export interface PaginatedResponse<T> {
+  data: T[]
+  total: number
+  skip: number
+  limit: number
+}
+
+export interface ApiError {
+  message: string
+  code: number
+  details?: unknown
+}
