+++ src/shared/lib/styles.ts
