+++ src/shared/ui/index.ts
@@ -1,0 +1,7 @@
+export { Button } from "./Button"
+export { Card } from "./Card"
+export { Dialog } from "./Dialog"
+export { Input } from "./Input"
+export { Select } from "./Select"
+export { Table } from "./Table"
+export { Textarea } from "./Textarea"
