+++ src/widgets/posts-manager/ui/index.ts
@@ -1,0 +1,1 @@
+export { PostsManagerWidget } from "./PostsManagerWidget"
