+++ vite.config.ts
@@ -1,8 +1,14 @@
 import react from "@vitejs/plugin-react"
+import path from "path"
 import { defineConfig } from "vite"
 
 export default defineConfig({
   plugins: [react()],
+  resolve: {
+    alias: {
+      "@": path.resolve(__dirname, "./src"),
+    },
+  },
   server: {
     proxy: {
       "/api": {
